@@ -3,31 +3,28 @@
 	"version": "0.0.0",
 	"private": true,
 	"engines": {
-		"node": "16.14.0"
+		"node": "20"
 	},
 	"scripts": {
 		"dev": "vite",
 		"build": "vite build",
 		"preview": "vite preview",
 		"test:unit": "vitest",
-		"prebuild": "node ./scripts/prebuild.js",
 		"lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
 	},
 	"dependencies": {
+		"@icholy/duration": "^5.1.0",
+		"@tanzanite/humanize-duration": "^4.0.1",
+		"axios": "^1.6.2",
 		"pinia": "^2.1.3",
 		"vue": "^3.3.4",
 		"vue-router": "^4.2.2"
 	},
 	"devDependencies": {
 		"@vitejs/plugin-vue": "^4.2.3",
-		"@vue/test-utils": "^2.3.2",
-		"axios": "^1.4.0",
-		"duration-js": "^4.0.0",
 		"eslint": "^8.39.0",
 		"eslint-plugin-vue": "^9.11.0",
-		"humanize-duration": "^3.28.0",
 		"jsdom": "^22.1.0",
-		"vite": "^4.3.9",
-		"vitest": "^0.32.0"
+		"vite": "^4.3.9"
 	}
 }
